double check ema

+++ mygui_docking/Project1/Project1/ADXL355/ADXL355Parser.h
@@ -8,12 +8,12 @@
 class ADXL355Parser {
 public:
     struct AccelerationData {
-        float x;  // X轴加速度 (g)
-        float y;  // Y轴加速度 (g)
-        float z;  // Z轴加速度 (g)
-        float EMA_x;
-        float EMA_y;
-        float EMA_z;
+        double x;  // X轴加速度 (g)
+        double y;  // Y轴加速度 (g)
+        double z;  // Z轴加速度 (g)
+        double EMA_x;
+        double EMA_y;
+        double EMA_z;
     };
 
     ADXL355Parser(uint8_t deviceId = 0x00) : deviceId_(deviceId) {}
